+++ CHANGELOG.md
@@ -11,9 +11,11 @@
     - devDependencies:
         - webpack-bundle-analyzer: 4.9.1
 - To manage dependencies of dependencies "overrides" for cesium are added to package.json and create dummy packages for it.
+- Issue #1105: Added easting and nothing labels for utm projections in CoordToolkit
 
 ### Changed
 - The script "npm run build" has been extended. The user can now specify the name of an addon that is excluded from bundle.
+- Issue #1064: Print: parameter to configure invisible layer info
 
 ### Deprecated